Add options to download and rewrite attachment urls
This adds option to provide a download location for message attachments.
If the option is set any received attachments will be automaticaly
downloaded to the specified location.

Additional to that there's an option to rewrite attachments urls. This
is useful in case you run a server that hosts downloaded files and want
to access the files without accessing slacks private file url.