Fix GetUsers and simplify GetUserInfo
Fixed GetUsers, it was missing a call to UserPagination.Failure, which
caused an infinite loop when getting users.

Also simplified GetUserInfo: there is no need to wrap this call around
a very slow function, when there is a more efficient one already.

Also locked GetUsers so it's not called concurrently, and made a few
minor cosmetic changes.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>