Disconnect client only if Slack disconnect is intentional
If the disconnection from Slack is intentiona, just keep closing the IRC
client connection as previously. However, if the disconnection is not
intentional, it's not necessary to disconnect the IRC client, as the
Slack reconnection will be handled by Slack's RTM.ManageConnection.
This should reduce the amount of disconnect/reconnect on the IRC client
side.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>