Change nick immediately on slack connect (#129)
If we wish to use the slack nickname in client communication, we must
first change the nick from the original submitted nick to the slack
nickname. Otherwise, if the slack nickname does not happen to match the
client-specified nickname, we just end up with wrong sources and
destinations for messages.

Store the original nickname for use in the rename process, reorder the
connect logic slightly to occur once we have all three needed
parameters, and make it so that the nickname is changed after slack
connect if it does not match the client specified one.